Initial drop of the front end pages.
Add yaml so that all of this serves up correctly.

Right now they don't do anything dynamic and make no calls.
Simple drag and drop functionality is there which is cool.

Made an edit to jquery-ui that allows a draggable to be moved
along a parabola. Very hacky. Doesn't take any parameters.